extract `bool_comparison` to a separate file
